fix(trading): glitchy waterfall loading (#5039)
