+++ apps/trading/components/bootstrapper/bootstrapper.tsx
@@ -4,11 +4,11 @@ import {
   AppLoader,
   DocsLinks,
   NetworkLoader,
+  NodeFailure,
   NodeGuard,
   useEnvironment,
 } from '@vegaprotocol/environment';
 import { t } from '@vegaprotocol/i18n';
-import { MaintenancePage } from '@vegaprotocol/ui-toolkit';
 import { VegaWalletProvider } from '@vegaprotocol/wallet';
 import type { ReactNode } from 'react';
 import { Web3Provider } from './web3-provider';
@@ -20,14 +20,9 @@ export const Bootstrapper = ({ children }: { children: ReactNode }) => {
     VEGA_ENV,
     VEGA_WALLET_URL,
     VEGA_EXPLORER_URL,
-    MAINTENANCE_PAGE,
     MOZILLA_EXTENSION_URL,
     CHROME_EXTENSION_URL,
   } = useEnvironment();
-
-  if (MAINTENANCE_PAGE) {
-    return <MaintenancePage />;
-  }
 
   if (
     !VEGA_URL ||
@@ -37,7 +32,7 @@ export const Bootstrapper = ({ children }: { children: ReactNode }) => {
     !MOZILLA_EXTENSION_URL ||
     !DocsLinks
   ) {
-    return null;
+    return <AppLoader />;
   }
 
   return (
@@ -50,9 +45,14 @@ export const Bootstrapper = ({ children }: { children: ReactNode }) => {
     >
       <NodeGuard
         skeleton={<AppLoader />}
-        failure={<AppFailure title={t(`Node: ${VEGA_URL} is unsuitable`)} />}
+        failure={<NodeFailure title={t(`Node: ${VEGA_URL} is unsuitable`)} />}
       >
-        <Web3Provider>
+        <Web3Provider
+          skeleton={<AppLoader />}
+          failure={
+            <AppFailure title={t(`Could not configure web3 provider`)} />
+          }
+        >
           <VegaWalletProvider
             config={{
               network: VEGA_ENV,
