+++ libs/environment/src/components/app-failure/app-failure.tsx
@@ -1,6 +1,4 @@
-import { t } from '@vegaprotocol/i18n';
-import { Button, Splash } from '@vegaprotocol/ui-toolkit';
-import { useNodeSwitcherStore } from '../../hooks/use-node-switcher-store';
+import { Splash } from '@vegaprotocol/ui-toolkit';
 
 export const AppFailure = ({
   title,
@@ -9,16 +7,10 @@ export const AppFailure = ({
   title: string;
   error?: string | null;
 }) => {
-  const setNodeSwitcher = useNodeSwitcherStore((store) => store.setDialogOpen);
   return (
     <Splash>
-      <div className="text-center">
-        <p className="mb-4 text-xl">{title}</p>
-        {error && <p className="mb-8 text-sm">{error}</p>}
-        <Button onClick={() => setNodeSwitcher(true)}>
-          {t('Change node')}
-        </Button>
-      </div>
+      <p className="mb-4 text-xl">{title}</p>
+      {error && <p className="mb-8 text-sm">{error}</p>}
     </Splash>
   );
 };
