+++ libs/environment/src/components/index.ts
@@ -4,4 +4,5 @@ export * from './network-loader';
 export * from './network-switcher';
 export * from './node-guard';
 export * from './node-switcher';
+export * from './node-failure';
 export * from './etherscan-link';
