+++ libs/environment/src/components/node-failure/index.ts
@@ -1,0 +1,1 @@
+export { NodeFailure } from './node-failure';
