--- libs/ui-toolkit/src/components/maintenance-page/index.ts
@@ -1,1 +1,0 @@
-export * from './maintenance-page';
